pyproject.toml
[project]
name = "road-defect"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cvzone>=1.6.1",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.8",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "ultralytics>=8.3.172",
]
